+++ library/ruler/src/test/java/io/mehow/ruler/DistanceSpec.kt
@@ -175,7 +175,7 @@ internal class DistanceSpec : BehaviorSpec({
             DistanceGenerator(Distance.ofKilometers(-1_000), Distance.ofKilometers(1_000)),
             Arb.long(0, 500_000)
         ) { distance, multiplicand ->
-          val meters = distance.exactTotalMeters * multiplicand.toBigDecimal()
+          val meters = distance.meters * multiplicand.toBigDecimal()
           val storedMeters = meters.toBigInteger().longValueExact()
           val nanometers = (meters - storedMeters.toBigDecimal()) * 1_000_000_000.toBigDecimal()
           val expected = Distance.create(storedMeters, nanometers.toLong())
@@ -193,7 +193,7 @@ internal class DistanceSpec : BehaviorSpec({
             DistanceGenerator(Distance.ofKilometers(-1_000), Distance.ofKilometers(1_000)),
             Arb.numericDoubles(0.0, 500_000.0)
         ) { distance, multiplicand ->
-          val meters = distance.exactTotalMeters * multiplicand.toBigDecimal()
+          val meters = distance.meters * multiplicand.toBigDecimal()
           val nanos = meters.movePointRight(9).toBigInteger()
           val divRem = nanos.divideAndRemainder(1_000_000_000.toBigInteger())
           check(divRem[0].bitLength() <= 63) { "Exceeded duration capacity: $nanos" }
@@ -212,7 +212,7 @@ internal class DistanceSpec : BehaviorSpec({
             DistanceGenerator(Distance.ofKilometers(-1_000), Distance.ofKilometers(1_000)),
             Arb.long(1, 500_000)
         ) { distance, divisor ->
-          val meters = distance.exactTotalMeters.divide(divisor.toBigDecimal(), DOWN)
+          val meters = distance.meters.divide(divisor.toBigDecimal(), DOWN)
           val nanos = meters.movePointRight(9).toBigIntegerExact()
           val divRem = nanos.divideAndRemainder(1_000_000_000.toBigInteger())
           check(divRem[0].bitLength() <= 63) { "Exceeded duration capacity: $nanos" }
@@ -231,7 +231,7 @@ internal class DistanceSpec : BehaviorSpec({
             DistanceGenerator(Distance.ofKilometers(-1_000), Distance.ofKilometers(1_000)),
             Arb.numericDoubles(0.000_001, 500_000.0)
         ) { distance, divisor ->
-          val meters = distance.exactTotalMeters.divide(divisor.toBigDecimal(), DOWN)
+          val meters = distance.meters.divide(divisor.toBigDecimal(), DOWN)
           val nanos = meters.movePointRight(9).toBigInteger()
           val divRem = nanos.divideAndRemainder(1_000_000_000.toBigInteger())
           check(divRem[0].bitLength() <= 63) { "Exceeded duration capacity: $nanos" }