+++ docs/changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `Distance` and `Length` throw now `ArithmeticException` in case of failures due to overflows or math operations.
+- Rename `exactTotalMeters` property to `meters`.
 - Upgrade to Kotlin `1.4.21`.
 
 ## [0.6.0] - 2020-11-11